Implemented purchase currency logic and creating the Order

+++ database/migrations/2022_11_09_180404_create_currencies_table.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->id();
             $table->string('code');
             $table->string('description');
-            $table->double('exchange_rate');
-            $table->double('surcharge');
-            $table->double('discount')->nullable();
+            $table->float('exchange_rate');
+            $table->float('surcharge');
+            $table->float('discount')->nullable();
         });
     }
